fixed cd

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: ahaloui <ahaloui@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/07 17:35:21 by aben-nei          #+#    #+#             */
-/*   Updated: 2023/06/04 21:41:00 by ahaloui          ###   ########.fr       */
+/*   Updated: 2023/06/04 22:07:41 by ahaloui          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include"libft.h"
 
-long	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	long		res;
 	int		sign;
